Split SimConfigs into FASEDTargetConfigs & CompilerConfigs
